fix(ent): restore schema fields removed in de1afa7 and regenerate ent code
- Add string ID fields to all ent schemas (Song, Playlist, Lyric, User, SyncLog, PlaylistSong)
- Restore missing fields: duration_ms, bitrate, sample_rate, owner_id (Song), cover_url, is_public, version (Playlist), offset_ms, source, version (Lyric), added_by (PlaylistSong)
- Rename user_id to owner_id in Playlist schema to match proto definition
- Remove invalid user_id index from Song schema
- Fix SongCount int/int32 type mismatch in convert.go and handler_playlist.go
- Fix SetCreatedAt -> SetAddedAt in playlist service.go
- Regenerate all ent ORM code with go generate

+++ internal/service/playlist/convert.go
@@ -19,7 +19,7 @@ func EntToProto(p *ent.Playlist) *playlistpb.Playlist {
 		CoverUrl:    p.CoverURL,
 		OwnerId:     p.OwnerID,
 		IsPublic:    p.IsPublic,
-		SongCount:   p.SongCount,
+		SongCount:   int32(p.SongCount),
 		Version:     p.Version,
 		CreatedAt:   convert.PTime(p.CreatedAt),
 		UpdatedAt:   convert.PTime(p.UpdatedAt),
